requirements.txt
aiohttp
pyyaml
py-cord==2.3.2